+++ spec/grape/api/instance_spec.rb
@@ -1,0 +1,48 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require 'shared/versioning_examples'
+
+describe Grape::API::Instance do
+  subject(:an_instance) do
+    Class.new(Grape::API::Instance) do
+      namespace :some_namespace do
+        get 'some_endpoint' do
+          'success'
+        end
+      end
+    end
+  end
+
+  let(:root_api) do
+    to_mount = an_instance
+    Class.new(Grape::API) do
+      mount to_mount
+    end
+  end
+
+  def app
+    root_api
+  end
+
+  context 'when an instance is mounted on the root' do
+    it 'can call the instance endpoint' do
+      get '/some_namespace/some_endpoint'
+      expect(last_response.body).to eq 'success'
+    end
+  end
+
+  context 'when an instance is the root' do
+    let(:root_api) do
+      to_mount = an_instance
+      Class.new(Grape::API::Instance) do
+        mount to_mount
+      end
+    end
+
+    it 'can call the instance endpoint' do
+      get '/some_namespace/some_endpoint'
+      expect(last_response.body).to eq 'success'
+    end
+  end
+end